server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
openai
python-docx
PyPDF2
pdf2image
pillow
pytesseract
